python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tikr"
version = "0.1.0"
description = "Python SDK for Tikr — schema-agnostic time series edge rollup engine"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [{ name = "walkerfunction" }]
keywords = ["timeseries", "grpc", "rollup", "hft", "telemetry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
]
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
]

[project.urls]
Homepage = "https://github.com/walkerfunction/tikr"
Repository = "https://github.com/walkerfunction/tikr"
Issues = "https://github.com/walkerfunction/tikr/issues"

[tool.setuptools.packages.find]
include = ["tikr*"]
